pyproject.toml
[project]
name = "de-project"
version = "0.1.0"
description = ""
authors = [
    {name = "timosii",email = "tsimashev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10, <3.14"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "dlt[postgres] (>=1.8.1,<2.0.0)",
    "pyarrow (>=19.0.1,<20.0.0)",
    "dbt-postgres (>=1.9.0,<2.0.0)",
    "superset (>=0.30.1,<0.31.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
